Move linked claim card code into claim details cards
