+++ app/presenters/payments/check_your_answers/ac_claim_details_card.rb
@@ -1,18 +1,10 @@
 module Payments
   module CheckYourAnswers
-    class AcClaimDetailsCard < BaseCard
-      attr_reader :session_answers, :params
-
-      def initialize(session_answers, params)
-        @session_answers = session_answers
-        @params = params
-
-        @section = 'claim_details'
-        super()
-      end
-
+    class AcClaimDetailsCard < ClaimDetailsCard
       def row_data
         [
+          linked_assigned_counsel,
+          linked_non_standard_magistrate,
           date_claim_assessed,
           solicitor_office_code,
           solicitor_firm_name,
@@ -21,6 +13,16 @@ module Payments
           ufn,
           defendant_last_name
         ].flatten.compact
+      end
+
+      def linked_assigned_counsel
+        return unless session_answers['request_type'].in?(%w[assigned_counsel_appeal assigned_counsel_amendment])
+
+        {
+          head_key: 'linked_assigned_counsel',
+          text: session_answers['laa_reference'] ||
+            I18n.t('payments.steps.check_your_answers.edit.sections.claim_details.no_linked_crm8')
+        }
       end
 
       def date_claim_assessed
